% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{amino_acid_order}
\alias{amino_acid_order}
\title{Standard amino acid ordering for heatmap y-axis}
\format{
An object of class \code{character} of length 23.
}
\usage{
amino_acid_order
}
\description{
A character vector defining the order of amino acids and special variant
types for the heatmap y-axis.
}
\keyword{datasets}
