% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\docType{data}
\name{classification_heatmap_colours}
\alias{classification_heatmap_colours}
\title{Classification colours for amino acid heatmap}
\format{
An object of class \code{character} of length 3.
}
\usage{
classification_heatmap_colours
}
\description{
A named character vector of colours for functional classification categories.
}
\keyword{datasets}
