% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{data}
\name{classification_shapes}
\alias{classification_shapes}
\title{Default shape palette for functional classifications}
\format{
A named numeric vector.
}
\usage{
classification_shapes
}
\description{
A named numeric vector mapping functional classifications to point shapes.
}
\examples{
classification_shapes

}
\keyword{datasets}
