% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{data}
\name{consequence_colours}
\alias{consequence_colours}
\title{Default color palette for consequence types}
\format{
A named character vector.
}
\usage{
consequence_colours
}
\description{
A named character vector mapping slim_consequence categories to colors.
}
\examples{
consequence_colours

}
\keyword{datasets}
