% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{data}
\name{default_consequences}
\alias{default_consequences}
\title{Default consequence categories for distribution plots}
\format{
An object of class \code{character} of length 11.
}
\usage{
default_consequences
}
\description{
A character vector of VEP slim consequence types shown by default in
ridgeline and jitter plots. The order determines the display order.
}
\keyword{datasets}
