% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgeasy-package.R
\docType{package}
\name{sgeasy-package}
\alias{sgeasy}
\alias{sgeasy-package}
\title{sgeasy: Downstream Analysis for Saturation Genome Editing Screens}
\description{
Provides tools for analyzing saturation genome editing (SGE) screen data,
including normalization using control oligos, differential abundance
analysis with DESeq2, and visualization utilities.
}
\section{Main functions}{

\describe{
\item{\code{\link{run_differential_analysis}}}{Run complete differential
abundance analysis pipeline}
\item{\code{\link{create_normalization_matrix}}}{Create matrix of neutral
variants for normalization}
\item{\code{\link{create_count_matrix}}}{Create count matrix from
annotated data}
\item{\code{\link{filter_by_counts}}}{Filter variants by count thresholds}
\item{\code{\link{estimate_size_factors}}}{Estimate DESeq2 size factors}
\item{\code{\link{plot_sample_scatter}}}{Create sample comparison plots}
\item{\code{\link{remove_artefacts}}}{Remove artefact variants}
\item{\code{\link{reweight_replicated_variants}}}{Combine replicated
variants using inverse variance weighting}
\item{\code{\link{sample_distance_matrix}}}{Create sample distance heatmap}
\item{\code{\link{build_sample_pca}}}{Build PCA plot from rlog data}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/team113sanger/sgeasy}
}

}
\author{
\strong{Maintainer}: Jamie Billington \email{jb63@sanger.ac.uk}

Authors:
\itemize{
  \item Andrew Waters \email{aw29@sanger.ac.uk}
}

}
\keyword{internal}
