% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SGEData}
\alias{SGEData}
\title{SGEData Class}
\usage{
SGEData(
  counts = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  metadata = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  annotation = NULL,
  complete_dataset = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })()
)
}
\arguments{
\item{counts}{A data frame of raw count data with SEQUENCE and SAMPLE columns.}

\item{metadata}{A data frame of sample metadata with condition information.}

\item{annotation}{A data frame of variant annotations (optional, can be NULL).}

\item{complete_dataset}{A data frame of counts joined with metadata.}
}
\value{
An S7 object of class SGEData.
}
\description{
An S7 class that holds loaded SGE screen data before analysis.
}
\details{
Access properties using the \code{@} operator:
\itemize{
\item \code{data@counts} - Raw count data
\item \code{data@metadata} - Sample metadata
\item \code{data@annotation} - Variant annotations (may be NULL)
\item \code{data@complete_dataset} - Counts joined with metadata
}
}
\examples{
\dontrun{
# Load data (returns SGEData object)
data <- load_sge_data(
  count_files = "inputs/",
  metadata_file = "metadata/screen_metadata.tsv"
)

# Print shows summary
data

# Access properties
data@counts
data@metadata
data@annotation
data@complete_dataset
}

}
