% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgedown-package.R
\docType{package}
\name{sgedown-package}
\alias{sgedown}
\alias{sgedown-package}
\title{sgedown: Downstream Analysis for Saturation Genome Editing Screens}
\description{
Provides tools for analyzing saturation genome editing (SGE) screen data,
including normalization using control oligos, differential abundance
analysis with DESeq2, and visualization utilities.
}
\section{Main functions}{

\describe{
\item{\code{\link{run_differential_analysis}}}{Run complete differential
abundance analysis pipeline}
\item{\code{\link{create_normalization_matrix}}}{Create matrix of neutral
variants for normalization}
\item{\code{\link{create_count_matrix}}}{Create count matrix from
annotated data}
\item{\code{\link{filter_by_counts}}}{Filter variants by count thresholds}
\item{\code{\link{estimate_size_factors}}}{Estimate DESeq2 size factors}
\item{\code{\link{plot_sample_scatter}}}{Create sample comparison plots}
}
}

\author{
\strong{Maintainer}: SGE Team \email{sge@sanger.ac.uk}

}
\keyword{internal}
