% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{slim_consequence}
\alias{slim_consequence}
\title{Simplify variant consequence annotations}
\usage{
slim_consequence(data)
}
\arguments{
\item{data}{A data frame with Consequence and mutator columns.}
}
\value{
Data frame with an additional slim_consequence column.
}
\description{
Creates a simplified consequence annotation column by mapping VEP consequences
to a reduced set of categories, considering mutator type where relevant.
}
\details{
The function maps VEP consequences to simplified categories:
\itemize{
\item Clinical inframe deletions/insertions (custom mutator)
\item Synonymous (including stop_retained with appropriate mutators)
\item Start/stop lost variants
\item Codon deletions (inframe mutator)
\item Stop gained, frameshift, splice variants
\item Missense, intron, UTR variants
}
}
\examples{
\dontrun{
data_with_slim <- slim_consequence(annotated_data)
}

}
